# [Type] Quat: small optimization for axisToQuat()