Avoiding doing unnecessary duplicate operations (norm and sin())
At first, I thought the compiler (msvc in my case) would be smart enough to do it by itself while optimizing, but apparently not:

before:
```
BM_Quat_axisToQuat/8192         106 us         85.4 us         8960
BM_Quat_axisToQuat/16384        210 us          154 us         4978
BM_Quat_axisToQuat/32768        348 us          220 us         3200
```

after:
```
BM_Quat_axisToQuat/8192        74.5 us         60.9 us        10000
BM_Quat_axisToQuat/16384        150 us         97.7 us         6400
BM_Quat_axisToQuat/32768        255 us          167 us         4480
```

----

Note: if we knew before-hand that the vec was normalized/not-null, the perf would be:
```
BM_Quat_axisToQuat/8192        69.0 us         52.7 us        16000
BM_Quat_axisToQuat/16384        138 us          103 us         7467
BM_Quat_axisToQuat/32768        239 us          171 us         4480
```

______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).