@alxbilger 2023-01-03
> At first, I thought the compiler (msvc in my case) would be smart enough to do it by itself while optimizing, but apparently not

Maybe because `phi` isn't const